rename file from SnpeLib to snpe_lib
