+++ onnxruntime/core/providers/snpe/snpe_execution_provider.h
@@ -3,7 +3,6 @@
 
 #pragma once
 
-#include "SnpeLib.h"
 #include <vector>
 #include <memory>
 #include <string>
