+++ onnxruntime/core/providers/snpe/snpe_runtime_options.cc
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#pragma once
-
 #include "core/providers/snpe/snpe_runtime_options.h"
 #include "core/common/logging/logging.h"
